+++ CHANGELOG.md
@@ -19,6 +19,31 @@ the following form: YYYY.0M.0D.
 - contains(String text), oldest(), pin(int index), and removeDuplicates().
 - Modified the interface to include a new kernel method, `clearEntry(int index)`
 - This made implement some of the secondary methods easier
+- Designed kernel and enhanced interfaces for Clipboard component
+
+### Updated
+
+- Changed design to include a new kernel method `entry(int index)` and added more
+- secondary methods such as `pin()` and `removeDuplicates()`
+
+## [2026.04.23]
+
+### Added
+
+- Designed a proof of concept for <Clipboard> component
+
+### Updated
+
+- Changed design to include kernel methods of
+  - copy
+  - paste
+  - clearEntry
+  - size
+  - isEmpty
+- As well as secondary methods of
+  - contains
+  - oldest
+- Also designed a main method that shows of the capabilities of the component
 
 ## [2026.04.23]